Adresses PR comments
